Refactored tn93 to separate module
